requirements.txt
absl-py==2.2.2
accelerate==1.10.1
aggdraw==1.3.19
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aiosqlite==0.21.0
alabaster==1.0.0
alembic==1.16.4
altair==5.5.0
annotated-types==0.7.0
anthropic==0.60.0
anyio==4.9.0
appdirs==1.4.4
appnope==0.1.4
asgiref==3.8.1
asttokens==3.0.0
astunparse==1.6.3
attrs==25.3.0
auth0-python==4.8.1
Automat==24.8.1
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.3
bitsandbytes==0.42.0
blinker==1.9.0
blis==1.2.0
bokeh==3.7.2
browserbase==1.4.0
bs4==0.0.2
build==1.2.2.post1
cachetools==5.5.2
catalogue==2.0.10
certifi==2025.1.31
cffi==1.17.1
chardet==3.0.4
charset-normalizer==3.4.1
chroma-hnswlib==0.7.6
chromadb==1.1.0
chromedriver==2.24.1
click==8.1.8
cloudpathlib==0.21.0
cohere==5.16.1
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.2
confection==0.1.5
constantly==23.10.4
contourpy==1.3.2
Crawl4AI==0.5.0.post4
crewai==0.152.0
crewai-tools==0.59.0
cryptography==44.0.2
cssselect==1.3.0
cv==1.0.0
cycler==0.12.1
cymem==2.0.11
dash==3.0.4
dash-core-components==2.0.0
dash-html-components==2.0.0
databases==0.9.0
dataclasses-json==0.6.7
datasets==4.1.1
debugpy==1.8.14
decorator==5.2.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
dill==0.4.0
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docstring_parser==0.16
docutils==0.21.2
duckduckgo_search==8.1.1
durationpy==0.9
embedchain==0.1.128
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
et_xmlfile==2.0.0
evaluate==0.4.6
executing==2.2.0
faiss-cpu==1.11.0.post1
fake-http-header==0.3.5
fake-useragent==2.1.0
Faker==37.3.0
fastapi==0.115.11
fastavro==1.12.0
faust-cchardet==2.1.19
filelock==3.18.0
Flask==3.0.3
flask-cors==4.0.0
flask-ngrok==0.0.25
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.5.0
fsspec==2025.3.0
gast==0.6.0
gensim==4.3.3
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.12
GitPython==3.1.44
google==3.0.0
google-auth==2.38.0
google-auth-oauthlib==1.2.2
google-pasta==0.2.0
google_search_results==2.4.2
googleapis-common-protos==1.69.2
googlesearch-python==1.3.0
googletrans==4.0.0rc1
gptcache==0.1.44
greenlet==3.1.1
groq==0.31.0
grpcio==1.71.0
gspread==6.2.1
h11==0.16.0
h2==3.2.0
h5py==3.13.0
hf-xet==1.1.5
hpack==3.0.0
hstspreload==2025.1.1
htmlmin==0.1.12
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
httpx-sse==0.4.0
huggingface-hub==0.34.3
humanfriendly==10.0
humanize==4.12.1
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
imagesize==1.4.1
imbalanced-learn==0.13.0
imblearn==0.0
importlib_metadata==8.6.1
importlib_resources==6.5.2
incremental==24.7.2
indic-nlp-library==0.92
iniconfig==2.1.0
instructor==1.7.7
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
json_repair==0.25.2
jsonpatch==1.33
jsonpickle==4.0.2
jsonpointer==3.0.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
keras==2.15.0
kiwisolver==1.4.8
kubernetes==32.0.1
lancedb==0.24.2
langchain==0.3.27
langchain-chroma==0.2.3
langchain-cohere==0.3.5
langchain-community==0.3.27
langchain-core==0.3.72
langchain-experimental==0.3.4
langchain-huggingface==0.3.1
langchain-openai==0.2.14
langchain-text-splitters==0.3.9
langcodes==3.5.0
langsmith==0.3.45
language_data==1.3.0
libclang==18.1.1
LiHiSTO==0.0.2
litellm==1.74.3
lxml==5.3.1
Mako==1.3.10
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdurl==0.1.2
mem0ai==0.1.115
ml-dtypes==0.3.2
mmh3==5.1.0
monotonic==1.6
Morfessor==2.0.6
mpmath==1.3.0
multidict==6.2.0
multiprocess==0.70.16
murmurhash==1.0.12
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.31.0
nest-asyncio==1.6.0
networkx==3.4.2
ngrok==1.4.0
nltk==3.9.1
nodeenv==1.9.1
numpy==1.26.4
oauth2client==4.1.3
oauthlib==3.2.2
onnxruntime==1.22.0
openai==1.98.0
opencv-python==4.11.0.86
openpyxl==3.1.5
opentelemetry-api==1.31.1
opentelemetry-exporter-otlp-proto-common==1.31.1
opentelemetry-exporter-otlp-proto-grpc==1.31.1
opentelemetry-exporter-otlp-proto-http==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-proto==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
opt_einsum==3.4.0
optree==0.15.0
orjson==3.10.15
outcome==1.3.0.post0
overrides==7.7.0
packaging==23.2
pandas==2.2.3
parsel==1.10.0
parso==0.8.4
pdfminer.six==20231228
pdfplumber==0.11.5
peft==0.17.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
playwright==1.51.0
plotly==6.0.1
pluggy==1.6.0
portalocker==2.7.0
posthog==3.21.0
preshed==3.0.9
primp==0.15.0
prompt_toolkit==3.0.51
propcache==0.3.0
Protego==0.4.0
protobuf==5.29.5
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py==1.11.0
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pybase64==1.4.2
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.10.1
pydantic_core==2.27.2
pydeck==0.9.1
PyDispatcher==2.0.7
pyee==12.1.1
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.11.2
pyngrok==7.2.4
pynput==1.8.1
pyobjc-core==11.0
pyobjc-framework-ApplicationServices==11.0
pyobjc-framework-Cocoa==11.0
pyobjc-framework-CoreText==11.0
pyobjc-framework-Quartz==11.0
pyOpenSSL==25.0.0
pyparsing==3.2.3
pypdf==5.9.0
pypdfium2==4.30.1
pyperclip==1.9.0
PyPika==0.48.9
pyproject_hooks==1.2.0
pyright==1.1.403
pysbd==0.3.4
PySocks==1.7.1
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
pytz==2025.2
pyvis==0.3.2
PyYAML==6.0.2
pyzmq==26.4.0
qdrant-client==1.15.1
queuelib==1.7.0
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
retry==0.9.2
retrying==1.3.4
rfc3986==1.5.0
rich==13.9.4
roman-numerals-py==3.1.0
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.3
schema==0.7.7
scikit-learn==1.6.1
scipy==1.13.1
scraper==0.1.0
scraperai==0.0.2
Scrapy==2.12.0
seaborn==0.13.2
selenium==4.9.1
sentence-transformers==3.4.1
serpapi==0.1.5
service-identity==24.2.0
shellingham==1.5.4
six==1.17.0
sklearn-compat==0.1.3
smart-open==7.1.0
smmap==5.0.2
smote==0.1
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
sounddevice==0.5.1
soupsieve==2.6
spacy==3.8.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
Sphinx==8.2.3
sphinx-argparse==0.5.2
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.39
srsly==2.5.1
stack-data==0.6.3
stagehand==0.5.0
starlette==0.46.1
streamlit==1.43.2
sympy==1.14.0
tabulate==0.9.0
tavily-python==0.7.10
tenacity==9.1.2
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.1
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.0.1
textblob==0.19.0
tf-playwright-stealth==1.1.2
tf_keras==2.20.1
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.9.0
tldextract==5.1.3
tokenizers==0.22.1
toml==0.10.2
tomli==2.2.1
tomli_w==1.2.0
torch==2.8.0
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.0
trio==0.29.0
trio-websocket==0.12.2
Twisted==24.11.0
typer==0.15.2
types-requests==2.32.4.20250611
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
undetected-chromedriver==3.5.5
urllib3==2.3.0
uv==0.6.9
uvicorn==0.34.0
uvloop==0.21.0
visualkeras==0.1.4
w3lib==2.3.1
wasabi==1.1.3
watchfiles==1.0.4
wcwidth==0.2.13
weasel==0.4.1
webdriver-manager==4.0.2
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.0.6
wrapt==1.14.2
wsproto==1.2.0
xxhash==3.5.0
xyzservices==2025.1.0
yarl==1.18.3
zipp==3.21.0
zope.interface==7.2
zstandard==0.23.0
